+++ src/main/java/com/sudo/railo/booking/infra/CartReservationRepository.java
@@ -1,4 +1,4 @@
-package com.sudo.railo.booking.infrastructure;
+package com.sudo.railo.booking.infra;
 
 import java.util.List;
